refactor: pkg/text → pkg/args/internal/text
