+++ pkg/args/args_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	. "github.com/celicoo/docli/pkg/args"
-	"github.com/celicoo/docli/pkg/text"
+	"github.com/celicoo/docli/pkg/args/internal/text"
 )
 
 const (
